Unify types deriving from atmp when needed.
