+++ src/types.ts
@@ -1,12 +1,4 @@
-import { Success, Error } from "./atmp/types.ts";
-
-/**
- * Items in the errors array returned by failed domain functions.
- */
-type ErrorWithMessage = {
-  message: string
-  exception?: unknown
-}
+import { Failure, Success } from './atmp/types.ts'
 
 /**
  * A successful domain function result.
@@ -19,7 +11,7 @@ type SuccessResult<T = void> = Success<T> & {
 /**
  * A failed domain function result.
  */
-type ErrorResult = Error & {
+type ErrorResult = Failure & {
   inputErrors: SchemaError[]
   environmentErrors: SchemaError[]
 }
@@ -102,60 +94,6 @@ type UnpackDFObject<Obj extends Record<string, DomainFunction>> =
   | never
 
 /**
- * Merges the data types of a list of objects.
- * @example
- * type MyObjs = [
- *   { a: string },
- *   { b: number },
- * ]
- * type MyData = MergeObjs<MyObjs>
- * //   ^? { a: string, b: number }
- */
-type MergeObjs<Objs extends unknown[], output = {}> = Objs extends [
-  infer first,
-  ...infer rest,
-]
-  ? MergeObjs<rest, Prettify<Omit<output, keyof first> & first>>
-  : output
-
-type Prettify<T> = {
-  [K in keyof T]: T[K]
-  // deno-lint-ignore ban-types
-} & {}
-
-/**
- * Converts a tuple type to a union type.
- * @example
- * type MyTuple = [string, number]
- * type MyUnion = TupleToUnion<MyTuple>
- * //   ^? string | number
- */
-type TupleToUnion<T extends unknown[]> = T[number]
-
-/**
- * Returns the last item of a tuple type.
- * @example
- * type MyTuple = [string, number]
- * type Result = Last<MyTuple>
- * //   ^? number
- */
-type Last<T extends readonly unknown[]> = T extends [...infer _I, infer L]
-  ? L
-  : never
-
-/**
- * It is similar to Partial<T> but it requires at least one property to be defined.
- * @example
- * type MyType = AtLeastOne<{ a: string, b: number }>
- * const a: MyType = { a: 'hello' }
- * const b: MyType = { b: 123 }
- * const c: MyType = { a: 'hello', b: 123 }
- * // The following won't compile:
- * const d: MyType = {}
- */
-type AtLeastOne<T, U = { [K in keyof T]: Pick<T, K> }> = Partial<T> & U[keyof U]
-
-/**
  * A parsing error when validating the input or environment schemas.
  * This will be transformed into a `SchemaError` before being returned from the domain function.
  * It is usually not visible to the end user unless one wants to write an adapter for a schema validator.
@@ -183,22 +121,25 @@ type ParserSchema<T extends unknown = unknown> = {
 
 export type {
   AtLeastOne,
-  DomainFunction,
-  ErrorData,
-  ErrorResult,
   ErrorWithMessage,
   Last,
   MergeObjs,
+  TupleToUnion,
+} from './atmp/types.ts'
+export type {
+  DomainFunction,
+  ErrorData,
+  ErrorResult,
   ParserIssue,
   ParserResult,
   ParserSchema,
   Result,
   SchemaError,
   SuccessResult,
-  TupleToUnion,
   UnpackAll,
   UnpackData,
   UnpackDFObject,
   UnpackResult,
   UnpackSuccess,
 }
+
